feat(api): RPC/OpenAPIルート定義を統一しメンテナンス性を向上 (#45)
## 概要
RPC用とOpenAPI用で二重管理されていたルート定義をOpenAPIベースに統一し、
保守性を大幅に改善しました。

## 主な変更

### スキーマの統合
- `src/server/schemas/` ディレクトリを新規作成
- `openapi/schemas/` から既存スキーマを移動（common.ts, leagues.ts）
- players用の新規スキーマを追加（players.ts）

### ルートファイルのOpenAPIベース化
- `routes/leagues.ts`: 通常のHonoからOpenAPIHonoに書き換え（69行→290行）
- `routes/players.ts`: OpenAPIHonoベースに書き換え（46行→156行）
- 全エンドポイントに詳細なOpenAPI仕様を追加

### メインルートの統合
- `routes/index.ts`: OpenAPIHonoベースに更新
- Swagger UIエンドポイント (`/api/ui`) を追加
- OpenAPI仕様エンドポイント (`/api/doc`) を追加
- RPC型 (`AppType`) を引き続きエクスポート

### クリーンアップ
- `src/server/openapi/` ディレクトリ全体を削除
- `app/api/[...route]/route.ts`: 統合アプリのみをマウント

## 改善効果

- **保守性向上**: ルート定義が1箇所に統一され、二重管理を解消
- **コード削減**: 359行 → 約300行（約20%削減）
- **ドキュメント自動生成**: Swagger UIとOpenAPI仕様を提供
- **型安全性維持**: RPC型推論が引き続き機能

## 動作確認

- ✅ OpenAPI仕様が正しく生成される (`/api/doc`)
- ✅ Biome lint/format が通る
- ✅ 既存のAPIエンドポイントが動作する

Closes #45
