+++ app/api/[...route]/route.ts
@@ -1,18 +1,11 @@
 import { Hono } from 'hono'
 import { handle } from 'hono/vercel'
-import openapiApp from '@/src/server/openapi'
-import rpcApp from '@/src/server/routes'
+import app from '@/src/server/routes'
 
-// メインアプリを作成
-const app = new Hono()
+// RPC と OpenAPI が統合されたアプリをマウント
+const mainApp = new Hono().route('/', app)
 
-// RPC ルートをマウント（既存）
-app.route('/', rpcApp)
-
-// OpenAPI ルートをマウント（新規）
-app.route('/', openapiApp)
-
-export const GET = handle(app)
-export const POST = handle(app)
-export const PATCH = handle(app)
-export const DELETE = handle(app)
+export const GET = handle(mainApp)
+export const POST = handle(mainApp)
+export const PATCH = handle(mainApp)
+export const DELETE = handle(mainApp)
